backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
defusedxml==0.7.1
distro==1.9.0
Django==5.2.5
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
git-filter-repo==2.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.99.6
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.4
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
youtube-transcript-api==1.2.2
